backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openpyxl==3.1.2
xlrd==2.0.1
werkzeug==3.0.1